\name{gamclass-package}
\alias{gamclass-package}
\alias{gamclass}
\docType{package}
\title{
Functions and Data for a Course in Modern Regression
}
\description{
For purposes of this package, modern regression extends to include
classification and multivariate exploration.  
A strong focus is on
methods described in Wood (2017) <doi:10.1201/9781315370279>
}
\details{
Functions are mostly designed to facilitate a variety of cross-validation
and bootstrap calculations.
}
\author{
John Maindonald

Maintainer: jhmaindonald@gmail.com
}
\references{
Venables, W N, & Ripley, B D (2013). Modern applied
statistics with S-PLUS. Springer Science & Business Media.

Wood, S N (2017) Generalized Additive Models: An Introduction
with R (2nd edition). Chapman and Hall/CRC.

\url{https://github.com/jhmaindonald/gamclass}
}
\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
