\name{modregR-package}
\alias{modregR-package}
\alias{modregR}
\docType{package}
\title{
Functions and Data for a Course in Modern Regression
}
\description{
For purposes of this package, modern regression extends to include
classification and multivariate exploration.  
A strong focus is on
methods described in Wood (2017) <doi:10.1201/9781315370279>
}
\details{
\tabular{ll}{
Package: \tab modregR\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2011-12-12\cr
License: \tab Unlimited \cr
}
Functions are mostly designed to facilitate a variety of cross-validation
and bootstrap calculations.
}
\author{
John Maindonald

Maintainer: john.maindonald@anu.edu.au
}
\references{
Venables, W N, & Ripley, B D (2013). Modern applied
statistics with S-PLUS. Springer Science & Business Media.

Wood, S N (2017) Generalized Additive Models: An Introduction
with R (2nd edition). Chapman and Hall/CRC.

\url{https://maths-people.anu.edu.au/~johnm/nzsr/taws.html}
}
\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
