\name{modregR-package}
\alias{modregR-package}
\alias{modregR}
\docType{package}
\title{
Functions and Data for a course in modern regression
}
\description{
For purposes of this package, modern regression extends to include
classification and multivariate exploration
}
\details{
\tabular{ll}{
Package: \tab modregR\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2011-12-12\cr
License: \tab Unlimited \cr
}
Functions are mostly designed to facilitate various cross-validation
and bootstrap calculations.
}
\author{
John Maindonald

Maintainer: john.maindonald@anu.edu.au
}
\references{
\url{http:/www.maths.anu.edu.au/~johnm/nzsr/taws.html}
}
\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
